+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes
 
 ## UNRELEASED
 
+* Feature: Add `misses` report to `--report` and log output.
 * Feature: Add `dynamic.resolutions` configuration option.
   [#4](https://github.com/FormidableLabs/trace-pkg/issues/4)